requirements.txt
numpy
scipy
argparse
Pillow